This is a working commit.
Created an interrupt controller.

Created a counter (so I could test the int ctlr)

Added instructions to cpu: ei, di, jsrint, reti.

Added func "Callback() bool" to cpu which is called on every
instruction boundary to check for interrupts.  full_sim populates
the callback with a func provided by the interrupt controller.

The interrupt controller has an array of 16 callbacks which default to nil.
By placing a callback (in full_sim) in the array, the int controller (on every Tick())
checks for interrupts. E.g. the counter provides a CounterIsZero() bool callback.
interruptController.Callbacks[0] = CounterIsZero() results in a (potential)
interrupt on every counter rollover.